Add files via upload
The Simplotel Assignment Web Page is a sleek and responsive design featuring a vibrant color palette and modern styling. The page includes a visually engaging image gallery with subtle image zoom on hover. A prominent card in the center showcases a bold red background, containing a title, Lorem Ipsum text, and a clickable button. The card exhibits a gentle zoom effect on hover, adding a dynamic touch. The overall design is both user-friendly and aesthetically pleasing, showcasing proficiency in HTML and CSS.